+++ src/VaccineRouter/model/VaccineRouter.h
@@ -13,12 +13,14 @@ private:
     Time vaccineLifeTime;
     std::string cityName;
     std::vector<StorageCenter *> SCs;
-    std::vector<ApplicationCenter *> ACs;
+    std::vector<ApplicationCenter *> selectedACs;
+    std::vector<ApplicationCenter *> availableACs;
 
     StorageCenter *findNearestSC();
     StorageCenter *findNearestSC(ApplicationCenter *applicationCenter);
     bool calculateSCRoute(StorageCenter *sc);
     void handleACsNotVisited();
+    void deleteDispatchedACs();
 public:
     VaccineRouter();
     VaccineRouter(Time vaccineLifeTime);
@@ -27,29 +29,27 @@ public:
     void setGraph(Graph *graph);
     void setCityName(std::string cityName);
 
-    int getCenter(Node *node);
     const std::vector<StorageCenter *> &getSCs() const;
-    const std::vector<ApplicationCenter *> &getACs() const;
+    const std::vector<ApplicationCenter *> &getSelectedACs() const;
+    const std::vector<ApplicationCenter *> &getAvailableACs() const;
+
+    int getCenter(Node *node);
+    std::string getCenterName(Node *node);
 
     void addStorageCenter(StorageCenter *sc);
+    void selectApplicationCenter(ApplicationCenter *ac);
     void addApplicationCenter(ApplicationCenter *ac);
 
     void selectMap(const std::string &mapFilename);
     bool setUpSCs(const std::string &mapFilename);
 
-    void processOrders();
     Time getVaccineLifeTime() const;
 
     void calculateRouteSingleSCSingleAC();
     void calculateRouteSingleSCMultipleAC();
     void calculateRouteSingleSCMultipleACWithTW();
     void calculateRouteMultipleSCMultipleACWithTW();
-
-    bool operator()(StorageCenter *sc);
-    void update();
-    void displayOutput();
-
-    std::string getCenterName(Node *node);
+    bool setUpACS(const std::string &basicString);
 };