+++ src/VaccineRouter/utilities/Coordinates.cpp
@@ -29,5 +29,5 @@ void Coordinates::setCoordinates(double x, double y) {
 }
 
 double Coordinates::calculateEuclideanDistance(Coordinates c) const{
-    return pow(std::sqrt(pow((this->x - c.getX()), 2) + pow((this->y - c.getY()), 2)), 2);
+    return std::sqrt(pow((this->x - c.getX()), 2) + pow((this->y - c.getY()), 2));
 }