+++ src/lib/GraphViewerCpp/SFML/extlibs/headers/stb_image/stb_image.h
@@ -1744,7 +1744,7 @@ static int stbi__build_huffman(stbi__huffman *h, int *count)
    return 1;
 }
 
-// build a table that decodes both magnitude and value of small ACs in
+// build a table that decodes both magnitude and value of small selectedACs in
 // one go.
 static void stbi__build_fast_ac(stbi__int16 *fast_ac, stbi__huffman *h)
 {